Fix field_offset! macro to avoid ptr -> int casts
These have been removed in recent nightlies,
since pointers can't nessicarrily have a meaningfull integer value.
I beleive this is because of issues with pointer 'provenance'.
Here is a good blog post explaining the issue in depth:
https://www.ralfj.de/blog/2020/12/14/provenance.html

+++ src/macros.rs
@@ -20,7 +20,7 @@ macro_rules! field_offset {
         unsafe {
             let uninit = core::mem::MaybeUninit::<$target>::uninit();
             let base = uninit.as_ptr();
-            (core::ptr::addr_of!((*base)$(.$field)*).cast::<u8>().offset_from(base as *const u8))
+            (core::ptr::addr_of!((*base)$(.$field)*).cast::<u8>().offset_from(base as *const u8) as usize)
         }
     }
 }
